[IMP] project_task_followers_mgmt_timesheet: user report
This commit adds a task analysis report taking in account that
in followers mode, more than a user is assigned to the task, then planned
and effective hours made in timesheets should be distributed between
these users.
This is achieved maintaning a "Timesheet summaries" structure that is later
used in reporting.