[FIX] project_task_followers_mgmt_timesheet: prevent timesheet summary access error
When accessing e.g. to OCA's Timesheet Sheets, an initial read to Timesheet
Summary is made for every user, raising an error for non-project managers.
With this fix read access is enabled for everyone. Since current frontend
access to Timesheet Summary in task form and Task Analysis menu is actually
restricted to Project Managers, no security leak is generated.